Moved a files arround for more scalable design
