+++ app/templates/edit_candidate.html
