+++ app/templates/electionprocess.html
