+++ app/templates/index.html
@@ -1,0 +1,266 @@
+<!DOCTYPE html>index
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>PALG UNN | Public Administration and Local Government</title>
+    <link
+      href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css"
+      rel="stylesheet"
+    />
+    <style>
+      body {
+        font-family: Arial, sans-serif;
+      }
+      .navbar {
+        background: #0056b3;
+      }
+      .navbar-brand img {
+        height: 50px;
+        width: auto;
+      }
+      .navbar-brand,
+      .nav-link {
+        color: #fff !important;
+      }
+      .carousel-item img {
+        width: 100%;
+        height: 500px;
+        object-fit: cover;
+      }
+      .hero-section {
+        height: 100vh;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        background: url("https://source.unsplash.com/1600x900/?university,students")
+          no-repeat center center/cover;
+        text-align: center;
+        color: white;
+        position: relative;
+      }
+      .hero-overlay {
+        position: absolute;
+        top: 0;
+        left: 0;
+        width: 100%;
+        height: 100%;
+        background: rgba(0, 86, 179, 0.7);
+        display: flex;
+        flex-direction: column;
+        justify-content: center;
+        align-items: center;
+        text-align: center;
+        padding: 20px;
+      }
+      .section {
+        padding: 80px 20px;
+      }
+      .section img {
+        width: 100%;
+        height: auto;
+        border-radius: 10px;
+      }
+      .footer {
+        background: #343a40;
+        color: white;
+        text-align: center;
+        padding: 20px;
+        font-size: 14px;
+      }
+      .contact-form {
+        background: #fff;
+        padding: 30px;
+        border-radius: 10px;
+        box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.1);
+      }
+    </style>
+  </head>
+  <body>
+    <nav class="navbar navbar-expand-lg">
+      <div class="container">
+        <a class="navbar-brand d-flex align-items-center" href="#">
+          <img src="../static/logo.png" alt="PALG Logo" class="me-2" />
+          <strong>PALG UNN</strong>
+        </a>
+        <button
+          class="navbar-toggler"
+          type="button"
+          data-bs-toggle="collapse"
+          data-bs-target="#navbarNav"
+        >
+          <span class="navbar-toggler-icon"></span>
+        </button>
+        <div class="collapse navbar-collapse" id="navbarNav">
+          <ul class="navbar-nav me-auto">
+            <li class="nav-item"><a class="nav-link" href="#home">Home</a></li>
+            <li class="nav-item">
+              <a class="nav-link" href="#about">About Us</a>
+            </li>
+            <li class="nav-item">
+              <a class="nav-link" href="#services">Services</a>
+            </li>
+            <li class="nav-item">
+              <a class="nav-link" href="#gallery">Gallery</a>
+            </li>
+            <li class="nav-item">
+              <a class="nav-link" href="#contact">Contact Us</a>
+            </li>
+          </ul>
+          <a href="{{ url_for('signup')}}" class="btn btn-light me-2">Signup</a>
+          <a href="{{ url_for('login')}}" class="btn btn-outline-light"
+            >Login</a
+          >
+        </div>
+      </div>
+    </nav>
+
+    <div
+      id="carouselExampleIndicators"
+      class="carousel slide"
+      data-bs-ride="carousel"
+    >
+      <div class="carousel-indicators">
+        <button
+          type="button"
+          data-bs-target="#carouselExampleIndicators"
+          data-bs-slide-to="0"
+          class="active"
+        ></button>
+        <button
+          type="button"
+          data-bs-target="#carouselExampleIndicators"
+          data-bs-slide-to="1"
+        ></button>
+        <button
+          type="button"
+          data-bs-target="#carouselExampleIndicators"
+          data-bs-slide-to="2"
+        ></button>
+      </div>
+      <div class="carousel-inner">
+        <div class="carousel-item active">
+          <img
+            src="../static/palg1.jpg"
+            class="d-block w-100"
+            alt="Campus Image"
+          />
+        </div>
+        <div class="carousel-item">
+          <img
+            src="../static/palg2.jpg"
+            class="d-block w-100"
+            alt="Lecture Image"
+          />
+        </div>
+        <div class="carousel-item">
+          <img
+            src="../static/palg3.jpg"
+            class="d-block w-100"
+            alt="Library Image"
+          />
+        </div>
+      </div>
+      <button
+        class="carousel-control-prev"
+        type="button"
+        data-bs-target="#carouselExampleIndicators"
+        data-bs-slide="prev"
+      >
+        <span class="carousel-control-prev-icon"></span>
+      </button>
+      <button
+        class="carousel-control-next"
+        type="button"
+        data-bs-target="#carouselExampleIndicators"
+        data-bs-slide="next"
+      >
+        <span class="carousel-control-next-icon"></span>
+      </button>
+    </div>
+
+    <div class="hero-section" id="home">
+      <div class="hero-overlay">
+        <h1>Welcome to PALG UNN</h1>
+        <p>Excellence in Public Administration and Local Government</p>
+      </div>
+    </div>
+
+    <div class="section" id="about">
+      <div class="container">
+        <h2>About Us</h2>
+        <p>
+          The Department of Public Administration and Local Government at UNN is
+          dedicated to shaping future leaders in governance.
+        </p>
+        <img
+          src="https://source.unsplash.com/800x500/?students,classroom"
+          alt="Department Image"
+        />
+      </div>
+    </div>
+
+    <div class="section bg-light" id="services">
+      <div class="container">
+        <h2>Our Services</h2>
+        <img
+          src="https://source.unsplash.com/800x500/?business,service"
+          alt="Services Image"
+        />
+      </div>
+    </div>
+
+    <div class="section" id="gallery">
+      <div class="container">
+        <h2>Gallery</h2>
+        <img
+          src="https://source.unsplash.com/800x500/?university,event"
+          alt="Gallery Image"
+        />
+      </div>
+    </div>
+
+    <div class="section bg-light" id="contact">
+      <div class="container">
+        <h2>Contact Us</h2>
+        <div class="row">
+          <div class="col-md-6">
+            <img
+              src="https://source.unsplash.com/800x500/?contact,communication"
+              alt="Contact Image"
+            />
+          </div>
+          <div class="col-md-6">
+            <div class="contact-form">
+              <h4>Get in Touch</h4>
+              <form>
+                <div class="mb-3">
+                  <label for="name" class="form-label">Full Name</label>
+                  <input type="text" class="form-control" id="name" required />
+                </div>
+                <div class="mb-3">
+                  <label for="email" class="form-label">Email Address</label>
+                  <input
+                    type="email"
+                    class="form-control"
+                    id="email"
+                    required
+                  />
+                </div>
+                <button type="submit" class="btn btn-primary w-100">
+                  Send Message
+                </button>
+              </form>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+
+    <footer class="footer">
+      <p>&copy; 2025 PALG UNN. All Rights Reserved.</p>
+    </footer>
+
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
+  </body>
+</html>
