--- templates/dashboard.html
@@ -1,216 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>lecturer admin Dashboard</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
-    <link rel="stylesheet" href="../static/assets/styles.css">
-    <style>
-        body {
-            font-family: 'Poppins', sans-serif;
-            background-color: #f8f9fa;
-        }
-
-        #wrapper {
-            display: flex;
-            flex-wrap: wrap;
-        }
-
-        /* Sidebar Styling */
-        #sidebar {
-            width: 250px;
-            min-height: 100vh;
-            background: #343a40;
-            color: white;
-            padding-top: 20px;
-            transition: all 0.3s;
-        }
-
-        #sidebar h2 {
-            font-size: 22px;
-            font-weight: bold;
-            text-align: center;
-        }
-
-        #sidebar ul {
-            padding-left: 0;
-        }
-
-        #sidebar ul li {
-            padding: 10px;
-            text-align: left;
-            transition: background 0.3s;
-        }
-
-        #sidebar ul li:hover {
-            background: #495057;
-            border-radius: 5px;
-        }
-
-        #sidebar ul li a {
-            text-decoration: none;
-            color: white;
-            display: block;
-        }
-
-        /* Page Content */
-        #page-content {
-            flex-grow: 1;
-            padding: 20px;
-            background: white;
-            border-radius: 8px;
-            box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.1);
-            margin: 20px;
-            width: 100%;
-        }
-
-        .profile-img {
-            width: 120px;
-            height: 120px;
-            border-radius: 50%;
-            object-fit: cover;
-            border: 3px solid #007bff;
-            display: block;
-            margin: 0 auto 10px;
-        }
-
-        /* Sidebar Toggle Button */
-        #sidebar-toggle {
-            display: none;
-            position: absolute;
-            top: 15px;
-            left: 15px;
-            background: #007bff;
-            color: white;
-            border: none;
-            padding: 8px 12px;
-            font-size: 16px;
-            cursor: pointer;
-            border-radius: 5px;
-        }
-
-        /* Responsive */
-        @media (max-width: 768px) {
-            #sidebar {
-                width: 100%;
-                min-height: auto;
-                display: none;
-                position: absolute;
-                top: 0;
-                left: 0;
-                z-index: 1000;
-            }
-
-            #sidebar-toggle {
-                display: block;
-            }
-
-            #wrapper {
-                flex-direction: column;
-            }
-        }
-    </style>
-</head>
-<body>
-    {% with messages = get_flashed_messages(with_categories=true) %}
-    {% if messages %}
-        {% for category, message in messages %}
-            <div class="alert alert-{{ category }}">{{ message }}</div>
-        {% endfor %}
-    {% endif %}
-{% endwith %}
-<!-- Sidebar Toggle Button -->
-<button id="sidebar-toggle">☰ Menu</button>
-
-<div class="d-flex" id="wrapper">
-    <!-- Sidebar -->
-    <div id="sidebar">
-        <h2>Dashboard</h2>
-        <ul class="list-unstyled">
-            <li><a href="{{ url_for('lecturer_admindashboard') }}">🏠 Home</a></li>
-            <li><a href="profile.html">👤 Profile</a></li>
-            <li><a href="courses.html">📚 Courses</a></li>
-            <li><a href="settings.html">⚙️ Settings</a></li>
-            <li><a href="{{ url_for('logout') }}">🚪 Logout</a></li>
-        </ul>
-    </div>
-  
-    <!-- Page Content -->
-    <div id="page-content">
-        <div class="container text-center">
-            <img src="../static/profile.webp" alt="Profile Picture" class="profile-img">
-            <h2>Welcome, {{ full_name }}! 👋</h2>
-            <p class="text-muted">You have successfully created an account.</p>
-        </div>
-
-        <!-- Seven Containers -->
-        <div class="container mt-4">
-            <div class="row">
-              <a href="{{ url_for('add_contestants') }}"> <div class="col-md-4 mb-4">
-                    <div class="p-4 bg-primary text-white rounded">
-                        <h4>Add Contestants</h4>
-                        <p>Students Going For Any Position...</p>
-                    </div>
-                </div>
-            </a> 
-                <div class="col-md-4 mb-4">
-                
-                    <a href="{{ url_for('add_students') }}"><div class="p-4 bg-secondary text-white rounded">
-                        <h4>Add Students</h4>
-                        <p>Add More Students...</p>
-                    </div>
-                </a> 
-                </div>
-                <div class="col-md-4 mb-4">
-                    <a href="{{ url_for('election_process') }}"> <div class="p-4 bg-success text-white rounded">
-                        <h4>Election</h4>
-                        <p>Election Process And Results...</p>
-                    </div>
-                </a> 
-
-                </div>
-
-                <div class="col-md-4 mb-4">
-                    <div class="p-4 bg-danger text-white rounded">
-                        <h4>Container 4</h4>
-                        <p>Content goes here...</p>
-                    </div>
-                </div>
-                <div class="col-md-4 mb-4">
-                    <div class="p-4 bg-warning text-dark rounded">
-                        <h4>Container 5</h4>
-                        <p>Content goes here...</p>
-                    </div>
-                </div>
-                <div class="col-md-4 mb-4">
-                    <div class="p-4 bg-info text-white rounded">
-                        <h4>Container 6</h4>
-                        <p>Content goes here...</p>
-                    </div>
-                </div>
-                <div class="col-md-12">
-                    <div class="p-4 bg-dark text-white rounded">
-                        <h4>Container 7</h4>
-                        <p>Content goes here...</p>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
-
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
-<script>
-    document.getElementById("sidebar-toggle").addEventListener("click", function() {
-        var sidebar = document.getElementById("sidebar");
-        if (sidebar.style.display === "block") {
-            sidebar.style.display = "none";
-        } else {
-            sidebar.style.display = "block";
-        }
-    });
-</script>
-
-</body>
-</html>
